Added HK2JunitRunner class to be used as an argument to @RunWith
