+++ hk2-testing/hk2-junitrunner/src/main/java/org/jvnet/hk2/testing/junit/HK2Runner.java
@@ -16,41 +16,15 @@
 
 package org.jvnet.hk2.testing.junit;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FilenameFilter;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.URL;
-import java.security.AccessController;
-import java.security.PrivilegedAction;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Enumeration;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.StringTokenizer;
-import java.util.jar.JarEntry;
-import java.util.jar.JarFile;
 
-import org.glassfish.hk2.api.DynamicConfiguration;
-import org.glassfish.hk2.api.DynamicConfigurationService;
 import org.glassfish.hk2.api.ServiceLocator;
-import org.glassfish.hk2.api.ServiceLocatorFactory;
-import org.glassfish.hk2.utilities.DescriptorImpl;
-import org.glassfish.hk2.utilities.ServiceLocatorUtilities;
 import org.junit.Before;
 import org.jvnet.hk2.testing.junit.annotations.Classes;
 import org.jvnet.hk2.testing.junit.annotations.Excludes;
-import org.jvnet.hk2.testing.junit.annotations.InhabitantFiles;
 import org.jvnet.hk2.testing.junit.annotations.Packages;
-import org.jvnet.hk2.testing.junit.internal.ClassVisitorImpl;
-import org.jvnet.hk2.testing.junit.internal.ErrorServiceImpl;
-import org.jvnet.hk2.testing.junit.internal.JustInTimeInjectionResolverImpl;
-import org.objectweb.asm.ClassReader;
+import org.jvnet.hk2.testing.junit.internal.TestServiceLocator;
 
 /**
  * This class should be extended by test classes in order to get an automatically
@@ -82,18 +56,16 @@ import org.objectweb.asm.ClassReader;
  * @author jwells
  */
 public class HK2Runner {
-    private final static String CLASS_PATH_PROP = "java.class.path";
-    private final static String DOT_CLASS = ".class";
+    private TestServiceLocator testServiceLocator;
     
     /**
      * Test classes can use this service locator as their private test locator
      */
     protected ServiceLocator testLocator;
-    
-    /**
-     * The verbosity of this runner
-     */
-    private boolean verbose = false;
+
+    public HK2Runner() {
+        testServiceLocator = new TestServiceLocator(this);
+    }
     
     /**
      * This will generate the default testLocator for this test
@@ -102,15 +74,8 @@ public class HK2Runner {
      */
     @Before
     public void before() {
-        Packages packages = getClass().getAnnotation(Packages.class);
-        if (packages == null) {
-            initialize(getClass().getName(), 
-                Collections.singletonList(getClass().getPackage().getName()),
-                null, null, null);
-        }
-        else {
-            initialize(null, null, null, null, null);
-        }
+        testServiceLocator.initializeOnBefore();
+        testLocator = testServiceLocator.getServiceLocator();
     }
     
     /**
@@ -120,7 +85,8 @@ public class HK2Runner {
      * {@link Packages}, {@link Classes}, {@link Excludes}, @{link InhabitantFiles}
      */
     public void initialize() {
-        initialize(null, null, null, null, null);
+        testServiceLocator.initialize();
+        testLocator = testServiceLocator.getServiceLocator();
     }
 
     /**
@@ -137,7 +103,8 @@ public class HK2Runner {
      * &#64;Service or not.  If null this is considered to be the empty set
      */
     protected void initialize(String name, List<String> packages, List<Class<?>> clazzes) {
-        initialize(name, packages, clazzes, null, null);
+        testServiceLocator.initialize(name, packages, clazzes);
+        testLocator = testServiceLocator.getServiceLocator();
     }
     
     /**
@@ -157,63 +124,8 @@ public class HK2Runner {
      * things coming from packages or from the hk2-locator/default file
      */
     protected void initialize(String name, List<String> packages, List<Class<?>> clazzes, Set<String> excludes) {
-        initialize(name, packages, clazzes, excludes, null);
-    }
-    
-    private List<String> getDefaultPackages() {
-        Packages packages = getClass().getAnnotation(Packages.class);
-        if (packages == null) return Collections.emptyList();
-        
-        List<String> retVal = new ArrayList<String>(packages.value().length);
-        for (String pack : packages.value()) {
-            if (Packages.THIS_PACKAGE.equals(pack)) {
-                retVal.add(getClass().getPackage().getName());
-            }
-            else {
-                retVal.add(pack);
-            }
-        }
-        
-        return retVal;
-    }
-    
-    private List<Class<?>> getDefaultClazzes() {
-        Classes clazzes = getClass().getAnnotation(Classes.class);
-        if (clazzes == null) return Collections.emptyList();
-        
-        List<Class<?>> retVal = new ArrayList<Class<?>>(clazzes.value().length);
-        for (Class<?> clazz : clazzes.value()) {
-            retVal.add(clazz);
-        }
-        
-        return retVal;
-    }
-    
-    private Set<String> getDefaultExcludes() {
-        Excludes excludes = getClass().getAnnotation(Excludes.class);
-        if (excludes == null) return Collections.emptySet();
-        
-        Set<String> retVal = new HashSet<String>();
-        for (String exclude : excludes.value()) {
-            retVal.add(exclude);
-        }
-        
-        return retVal;
-    }
-    
-    private Set<String> getDefaultLocatorFiles() {
-        HashSet<String> retVal = new HashSet<String>();
-        InhabitantFiles iFiles = getClass().getAnnotation(InhabitantFiles.class);
-        if (iFiles == null) {
-            retVal.add("META-INF/hk2-locator/default");
-            return retVal;
-        }
-        
-        for (String iFile : iFiles.value()) {
-            retVal.add(iFile);
-        }
-        
-        return retVal;
+        testServiceLocator.initialize(name, packages, clazzes, excludes);
+        testLocator = testServiceLocator.getServiceLocator();
     }
     
     /**
@@ -237,251 +149,11 @@ public class HK2Runner {
      * values those will be searched as resources from the jars in the classpath to load the registry with
      */
     protected void initialize(String name, List<String> packages, List<Class<?>> clazzes, Set<String> excludes, Set<String> locatorFiles) {
-        if (name == null) name = getClass().getName();
-        if (packages == null) packages = getDefaultPackages();
-        if (clazzes == null) clazzes = getDefaultClazzes();
-        if (excludes == null) excludes = getDefaultExcludes();
-        if (locatorFiles == null) locatorFiles = getDefaultLocatorFiles();
-        
-        ServiceLocator found = ServiceLocatorFactory.getInstance().find(name);
-        if (found != null) {
-            testLocator = found;
-            
-            testLocator.inject(this);
-            return;
-        }
-        
-        testLocator = ServiceLocatorFactory.getInstance().create(name);
-        
-        ServiceLocatorUtilities.addClasses(testLocator, ErrorServiceImpl.class);
-        final JustInTimeInjectionResolverImpl jitResolver = new JustInTimeInjectionResolverImpl(excludes);
-        testLocator.inject(jitResolver);
-        ServiceLocatorUtilities.addOneConstant(testLocator, jitResolver);
-        
-        DynamicConfigurationService dcs = testLocator.getService(DynamicConfigurationService.class);
-        DynamicConfiguration config = dcs.createDynamicConfiguration();
-        
-        addServicesFromDefault(config, excludes, locatorFiles);
-        
-        addServicesFromPackage(config, packages, excludes);
-        
-        for (Class<?> clazz : clazzes) {
-            config.addActiveDescriptor(clazz);
-        }
-        
-        config.commit();
-        
-        testLocator.inject(this);
+        testServiceLocator.initialize(name, packages, clazzes, excludes, locatorFiles);
+        testLocator = testServiceLocator.getServiceLocator();
     }
     
     protected void setVerbosity(boolean verbose) {
-        this.verbose = verbose;
-    }
-    
-    private void addServicesFromDefault(final DynamicConfiguration config, final Set<String> excludes, final Set<String> locatorFiles) {
-        AccessController.doPrivileged(new PrivilegedAction<Object>() {
-
-            @Override
-            public Object run() {
-                internalAddServicesFromDefault(config, excludes, locatorFiles);
-                return null;
-            }
-            
-        });
-    }
-    
-    private void readResources(Enumeration<URL> resources, Set<String> excludes, DynamicConfiguration config) {
-        while (resources.hasMoreElements()) {
-            URL url = resources.nextElement();
-           
-            try {
-                InputStream urlStream = url.openStream();
-                
-                BufferedReader reader = new BufferedReader(new InputStreamReader(urlStream));
-                
-                boolean goOn = true;
-                while (goOn) {
-                    DescriptorImpl bindMe = new DescriptorImpl();
-                
-                    goOn = bindMe.readObject(reader);
-                    if (goOn == true && !excludes.contains(bindMe.getImplementation())) {
-                        config.bind(bindMe);
-                    }
-                }
-                
-                reader.close();
-            }
-            catch (IOException ioe) {
-                ioe.printStackTrace();
-                
-                continue;
-            }
-            
-            
-        }
-        
-    }
-    
-    private void internalAddServicesFromDefault(DynamicConfiguration config, Set<String> excludes, Set<String> locatorFiles) {
-        ClassLoader loader = this.getClass().getClassLoader();
-        
-        for (String locatorFile : locatorFiles) {
-            Enumeration<URL> resources;
-            try {
-                resources = loader.getResources(locatorFile);
-            }
-            catch (IOException ioe) {
-                ioe.printStackTrace();
-            
-                return;
-            }
-            
-            readResources(resources, excludes, config);
-        }
-    }
-    
-    private void addServicesFromPackage(final DynamicConfiguration config, final List<String> packages, final Set<String> excludes) {
-        AccessController.doPrivileged(new PrivilegedAction<Object>() {
-
-            @Override
-            public Object run() {
-                internalAddServicesFromPackage(config, packages, excludes);
-                return null;
-            }
-            
-        });
-    }
-    
-    private void internalAddServicesFromPackage(DynamicConfiguration config, List<String> packages, Set<String> excludes) {
-        if (packages.isEmpty()) {
-            return;
-        }
-        
-        String classPath = AccessController.doPrivileged(new PrivilegedAction<String>() {
-            @Override
-            public String run() {
-                return System.getProperty(CLASS_PATH_PROP);
-            }
-                
-        });
-        
-        StringTokenizer st = new StringTokenizer(classPath, File.pathSeparator);
-        
-        while(st.hasMoreTokens()) {
-            String pathElement = st.nextToken();
-            
-            addServicesFromPathElement(config, packages, pathElement, excludes);
-        }
-        
-    }
-    
-    private void addServicesFromPathElement(DynamicConfiguration config, List<String> packages, String element, Set<String> excludes) {
-        File fileElement = new File(element);
-        if (!fileElement.exists()) return;
-        
-        if (fileElement.isDirectory()) {
-            addServicesFromPathDirectory(config, packages, fileElement, excludes);
-        }
-        else {
-            addServicesFromPathJar(config, packages, fileElement, excludes);
-        }
-    }
-    
-    private void addServicesFromPathDirectory(DynamicConfiguration config, List<String> packages, File directory, Set<String> excludes) {
-        for (String pack : packages) {
-            File searchDir = new File(directory, convertToFileFormat(pack));
-            if (!searchDir.exists()) continue;
-            if (!searchDir.isDirectory()) continue;
-            
-            File candidates[] = searchDir.listFiles(new FilenameFilter() {
-
-                @Override
-                public boolean accept(File dir, String name) {
-                    if (name == null) return false;
-                    if (name.endsWith(DOT_CLASS)) return true;
-                    return false;
-                }
-                
-            });
-            
-            if (candidates == null) continue;
-            
-            for (File candidate : candidates) {
-                try {
-                    FileInputStream fis = new FileInputStream(candidate);
-                    
-                    addClassIfService(fis, excludes);
-                }
-                catch (IOException ioe) {
-                    // Just don't add it
-                }
-            }
-        }
-        
-    }
-    
-    private void addServicesFromPathJar(DynamicConfiguration config, List<String> packages, File jar, Set<String> excludes) {
-        JarFile jarFile;
-        try {
-            jarFile = new JarFile(jar);
-        }
-        catch (IOException ioe) {
-            // Not a jar file, forget it
-            return;
-        }
-        
-        try {
-            for (String pack : packages) {
-                String packAsFile = convertToFileFormat(pack);
-                int packAsFileLen = packAsFile.length() + 1;
-            
-                Enumeration<JarEntry> entries = jarFile.entries();
-                while (entries.hasMoreElements()) {
-                    JarEntry entry = entries.nextElement();
-                
-                    String entryName = entry.getName();
-                    if (!entryName.startsWith(packAsFile)) {
-                        // Not in the correct directory
-                        continue;
-                    }
-                    if (entryName.substring(packAsFileLen).contains("/")) {
-                        // Next directory down
-                        continue;
-                    }
-                    if (!entryName.endsWith(DOT_CLASS)) {
-                        // Not a class
-                        continue;
-                    }
-                
-                    try {
-                        addClassIfService(jarFile.getInputStream(entry), excludes);
-                    }
-                    catch (IOException ioe) {
-                        // Simply don't add it if we can't read it
-                    }
-                }
-            }
-        }
-        finally {
-            try {
-                jarFile.close();
-            }
-            catch (IOException e) {
-                // Ignore
-            }
-        }
-    }
-    
-    private void addClassIfService(InputStream is, Set<String> excludes) throws IOException {
-        ClassReader reader = new ClassReader(is);
-        
-        ClassVisitorImpl cvi = new ClassVisitorImpl(testLocator, verbose, excludes);
-        
-        reader.accept(cvi, ClassReader.SKIP_CODE | ClassReader.SKIP_DEBUG | ClassReader.SKIP_FRAMES);
-        
-    }
-    
-    private static String convertToFileFormat(String clazzFormat) {
-        return clazzFormat.replaceAll("\\.", "/");
+        testServiceLocator.setVerbosity(verbose);
     }
 }
